Create OldProcessingManager
